+++ packages/docs/src/data/api/rules/email.ts
@@ -1,0 +1,12 @@
+import { Api } from '~/types';
+import { ruleMessages } from '../shared/ruleMessages';
+
+export const api: Api = {
+	email: {
+		arguments: [
+			...ruleMessages(`{
+	default: 'L\\'email saisi est invalide.'
+}`)
+		]
+	}
+};